requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
sqlparse~=0.4.3
pip install django-taggit==3.0.0
pip install markdown==3.4.1
pip install psycopg2-binary==2.9.3
pip~=23.1
keyring~=23.13.1
colorama~=0.4.3
pytz~=2019.3
asyncio~=3.4.3
PyYAML~=5.3.1
requests~=2.22.0
djangorestframework~=3.14.0
Pygments~=2.3.1
packaging~=23.1
future~=0.18.2
pexpect~=4.8.0
setuptools~=67.6.1
PyGObject~=3.36.0
simplejson~=3.16.0
ptyprocess~=0.7.0
Markdown~=3.4.3
Pillow~=7.0.0
six~=1.14.0
zipp~=3.15.0
twilio~=8.0.0
python-dotenv==0.21.0